Removal Request, UW Homework Solutions
Hello, this is a trivial change only made so that I could send this message. I work with the University of Washington on CSE 142 and the homework assignment solutions in this repo are getting copied by our students frequently. Would you be willing to delete this repo to help us avoid misconduct and plagiarism in our courses?